Test boolean options for nil in program conf templates
The old behavior would elide the option from the file if the parm
was explicitly set to false. This doesn't work if the default value
of the parm is true (autostart, for example)

The new behavior properly distinguishes undef from false, and will
insert an explicit 'option = false' into the program conf file.

For details, see:

  https://docs.puppetlabs.com/guides/templating.html#testing-for-undefined-variables

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>