Allowing NONE logging; Checking for undefined boolean variables
Supervisord allows special values NONE and AUTO for the stdout
and stderr log paths. These should not be prefixed with the
path of the logging directory, which is ensured by this commit.

All boolean values in supervisord::program are also checked
against null, rather than false. Therefore, the user is able
to specifically add 'false' value to the config.