Adding support for proxy
In cases where node does not have direct internet connection, a proxy is required. Since older version of pip does not honor proxy configuration setting the simplest way is to use exec with environment.
We tried also setting the proxy in puppet.conf file but it has the unwanted effect of using the proxy for puppet agent-master communication.